Rename openweathermap coordinator module
