+++ .coveragerc
@@ -979,9 +979,9 @@ omit =
     homeassistant/components/openuv/coordinator.py
     homeassistant/components/openuv/sensor.py
     homeassistant/components/openweathermap/__init__.py
+    homeassistant/components/openweathermap/coordinator.py
     homeassistant/components/openweathermap/sensor.py
     homeassistant/components/openweathermap/weather.py
-    homeassistant/components/openweathermap/weather_update_coordinator.py
     homeassistant/components/opnsense/__init__.py
     homeassistant/components/opnsense/device_tracker.py
     homeassistant/components/opower/__init__.py
