+++ homeassistant/components/openweathermap/__init__.py
@@ -26,7 +26,7 @@ from .const import (
     FORECAST_MODE_ONECALL_DAILY,
     PLATFORMS,
 )
-from .weather_update_coordinator import WeatherUpdateCoordinator
+from .coordinator import WeatherUpdateCoordinator
 
 _LOGGER = logging.getLogger(__name__)
